+++ crates/nimbis/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cli;
 pub mod client;
 pub mod cmd;
 pub mod config;
